+++ src/algorithms/objects/image.cpp
@@ -1,6 +1,6 @@
 #include "image.h"
 
-image::image(int width, int height)
+image::image(unsigned long width, unsigned long height)
 {
     this->_width = width;
     this->_height = height;
